+++ Eyediagram/main/run.py
@@ -1,4 +1,5 @@
-import sys, os, glob, set
+import sys, os, glob
+from src import set
 sys.path.append(os.path.dirname(os.path.abspath(os.path.dirname(__file__))))
 
 from src import Eye
